Changes bases on PR-692 feedback; moved most ingestion modules to the ingestion sub-directory.
