+++ src/encoded/__init__.py
@@ -193,8 +193,8 @@ def main(global_config, **local_config):
     config.include('.loadxl')
     config.include('.visualization')
     config.include('.ingestion_listener')
-    config.include('.ingestion_message_handler_vcf')
-    config.include('.ingestion_message_handler_default')
+    config.include('.ingestion.ingestion_message_handler_vcf')
+    config.include('.ingestion.ingestion_message_handler_default')
     config.include('.custom_embed')
 
     if 'elasticsearch.server' in config.registry.settings:
