+++ src/encoded/ingestion_listener.py
@@ -26,6 +26,13 @@ from pyramid.view import view_config
 from snovault.util import debug_log
 from .ingestion.common import metadata_bundles_bucket, get_parameter, IngestionReport
 from .ingestion.exceptions import UnspecifiedFormParameter, SubmissionFailure  # , BadParameter
+from .ingestion.ingestion_listener_base import (
+    STATUS_QUEUED,
+    STATUS_INGESTED,
+    DEBUG_SUBMISSIONS,
+    IngestionListenerBase,
+)
+from .ingestion.ingestion_message_handler_decorator import call_ingestion_message_handler
 from .ingestion.processors import get_ingestion_processor
 from .ingestion.queue_utils import IngestionQueueManager
 from .types.ingestion import SubmissionFolio, IngestionSubmission
@@ -34,13 +41,6 @@ from .util import (
     subrequest_object, register_path_content_type, vapp_for_email,  # vapp_for_ingestion,
     SettingsKey, make_s3_client, extra_kwargs_for_s3_encrypt_key_id,
 )
-from .ingestion_listener_base import (
-    STATUS_QUEUED,
-    STATUS_INGESTED,
-    DEBUG_SUBMISSIONS,
-    IngestionListenerBase,
-)
-from .ingestion_message_handler_decorator import call_ingestion_message_handler
 
 
 log = structlog.getLogger(__name__)
